Rename module again
