+++ run_tests.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-python -m unittest discover -s pathpipeconftools/tests/ -p '*_test.py'
+python -m unittest discover -s path_pipe_conf_tools/tests/ -p '*_test.py'
